backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
redis>=5.0.0
mysql-connector-python>=8.2.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.25.0
requests>=2.31.0
